requirements.txt
numpy==1.19.5
pandas==1.1.5
tensorflow-gpu==1.15.4